separate tendermint wrapper funcs
